refactored code for tooltip
